requirements.txt
requests==2.26.0
kafka-python==2.0.2
psycopg2==2.9.1
python-dotenv==0.19.1
python-dateutil==2.8.2